+++ Changelog.md
@@ -17,6 +17,8 @@ This file contains the changelog of SwiftExtras.
   A method that takes a snapshot of the view and returns a platform image
 - Added [`View.snapshot(name:size:)`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/swiftuicore/view/snapshot(name:size:))  
   A method that takes a snapshot of the view and returns a platform image
+- Added `priority` parameter to [`View.detachedTask(priority:_:)`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/swiftuicore/view/detachedtask(priority:_:))  
+  A view modifier that performs a detached task
 
 ### 1.2.5
 - Updated [`CardView`](https://0xwdg.github.io/SwiftExtras/documentation/swiftextras/cardview)